r691 update CPMlib 103

+++ src/CPMlib-1.0.3/include/cpm_Define.h
@@ -54,9 +54,9 @@
  *  @return 1次元インデクス
  */
 #define _IDX_S3D(_I,_J,_K,_NI,_NJ,_NK,_VC) \
-( size_t(_K+_VC) * size_t(_NI+2*_VC) * size_t(_NJ+2*_VC) \
-+ size_t(_J+_VC) * size_t(_NI+2*_VC) \
-+ size_t(_I+_VC) \
+( (long long)(_K+_VC) * (long long)(_NI+2*_VC) * (long long)(_NJ+2*_VC) \
++ (long long)(_J+_VC) * (long long)(_NI+2*_VC) \
++ (long long)(_I+_VC) \
 )
 
 /** 4次元インデクス(i,j,k,n) -> 1次元インデクス変換マクロ
@@ -71,7 +71,7 @@
  *  @return 1次元インデクス
  */
 #define _IDX_S4D(_I,_J,_K,_N,_NI,_NJ,_NK,_VC) \
-( size_t(_N) * size_t(_NI+2*_VC) * size_t(_NJ+2*_VC) * size_t(_NK+2*_VC) \
+( (long long)(_N) * (long long)(_NI+2*_VC) * (long long)(_NJ+2*_VC) * (long long)(_NK+2*_VC) \
 + _IDX_S3D(_I,_J,_K,_NI,_NJ,_NK,_VC) \
 )
 
@@ -100,8 +100,8 @@
  *  @return 1次元インデクス
  */
 #define _IDX_S4DEX(_N,_I,_J,_K,_NN,_NI,_NJ,_NK,_VC) \
-( size_t(_NN) * _IDX_S3D(_I,_J,_K,_NI,_NJ,_NK,_VC) \
-+ size_t(_N) )
+( (long long)(_NN) * _IDX_S3D(_I,_J,_K,_NI,_NJ,_NK,_VC) \
++ (long long)(_N) )
 
 /** 3次元インデクス(3,i,j,k) -> 1次元インデクス変換マクロ
  *  @param[in] _N  成分インデクス
